backend/setup/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework.authtoken import views

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('wall.urls')),
    path('api/auth/', include('accounts.urls')),
]
