backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
python-dotenv==1.0.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
